feat: Include a button on the Navigation screen to make the view follow the ship's position. Closes #386
