+++ client/src/components/Starmap/useCancelFollow.tsx
@@ -1,0 +1,17 @@
+import {useGetStarmapStore} from "client/src/components/Starmap/starmapStore";
+import {useEffect} from "react";
+
+export function useCancelFollow() {
+  const useStarmapStore = useGetStarmapStore();
+  const {cameraControls} = useStarmapStore();
+  useEffect(() => {
+    const cancel = () => {
+      console.log("Cancel");
+      useStarmapStore.setState({followEntityId: null});
+    };
+    cameraControls?.current?.addEventListener("controlstart", cancel);
+    return () => {
+      cameraControls?.current?.removeEventListener("controlstart", cancel);
+    };
+  });
+}
